refactor : 통계 부분 파일 분리
