+++ srcs/problem/problem.controller.js
@@ -2,7 +2,6 @@ import { status } from "../../config/response.status.js";
 import { ProblemService } from "./problem.service.js";
 import { ChatService } from "../chat/chat.service.js";
 import { response } from "../../config/response.js";
-import { createMulter, getPublicUrl } from "../utils/image/image.upload.js";
 import { 
   setScaleResponseDTO, 
   getProblemListResponseDTO, 
@@ -13,9 +12,6 @@ import {
   problemTypeResponseDTO ,
   addProblemTypeResponseDTO,
   deleteProblemResponseDTO,
-  getStatisticIncorrectProblemDTO,
-  getStatisticIncorrectTypeDTO,
-  getStatisticIncorrectRatioDTO, 
 } from "./problem.reponse.dto.js";
 
 export const setScale = async (req, res) => {
@@ -282,59 +278,3 @@ export const deleteProblemType = async (req, res) => {
     res.send(response(status.INTERNAL_SERVER_ERROR));
   }
 };
-
-// 가장 많이 틀린 문제 가져오기
-export const getStatisticIncorrectProblem = async (req, res) => {
-    try{
-    const userId = req.userId;
-    const statistic = await ProblemService.getStatisticIncorrectProblem(userId);
-
-    res.send(response(status.PROBLEM_STATISTIC_SUCCESS,statistic));
-    //res.send(response(status.PROBLEM_STATISTIC_SUCCESS,getStatisticIncorrectProblemDTO(statistic)));
-
-    } catch (error){
-      res.send(response(status.INTERNAL_SERVER_ERROR));
-  }
-}
-
-export const getStatisticIncorrectType = async (req, res) => {
-  try {
-      const userId = req.userId;
-      const statistic = await ProblemService.getStatisticIncorrectType(userId);
-      
-      if (!statistic) {
-          return res.send(response(status.NO_CONTENT, null));  // 틀린 문제가 없는 경우 처리
-      }
-
-      // DTO로 변환하여 응답
-      const statisticDTO = getStatisticIncorrectTypeDTO(statistic);
-      res.send(response(status.PROBLEM_STATISTIC_SUCCESS, statisticDTO));
-  } catch (error) {
-      res.send(response(status.INTERNAL_SERVER_ERROR));
-  }
-}
-
-
-
-export const getStatisticIncorrectRatio = async (req, res) => {
-  try {
-      const userId = req.userId;
-      const { categoryId } = req.params;  // category_id를 파라미터로 받음
-
-      if (!categoryId) {
-          return res.send(response(status.BAD_REQUEST, "카테고리 ID가 필요합니다."));
-      }
-
-      const statistic = await ProblemService.getStatisticIncorrectRatio(userId, categoryId);
-
-      if (!statistic || statistic.length === 0) {
-          return res.send(response(status.NO_CONTENT, null));  // 데이터가 없는 경우 처리
-      }
-
-      // DTO로 변환하여 응답
-      const statisticDTO = getStatisticIncorrectRatioDTO(statistic);
-      res.send(response(status.PROBLEM_STATISTIC_SUCCESS, statisticDTO));
-  } catch (error) {
-      res.send(response(status.INTERNAL_SERVER_ERROR));
-  }
-}
