+++ srcs/problem/problem.service.js
@@ -162,7 +162,6 @@ export const ProblemService = {
       } = problemData;
       await ProblemModel.updateProblem(problemId, problemText, answer, status, memo);
       console.log('Service - updateProblem: 문제 수정 완료');
-      console.log(photos)
       if (photos) {
         const photoTypes = Object.keys(photos);
         for (const photoType of photoTypes) {
@@ -299,56 +298,4 @@ export const ProblemService = {
       throw new BaseError(status.BAD_REQUEST, "소분류 삭제 실패");
     }
   },
-
-  getStatisticIncorrectProblem: async (userId) => {
-    try{
-      
-      return await ProblemModel.getIncorrectProblemStatistic(userId);
-    } catch (error) {
-      throw new BaseError(status.INTERNAL_SERVER_ERROR,"틀린 문제 통계 조회 실패")
-    }
-  },
-  
-
-  getStatisticIncorrectType: async (userId) => {
-    try {
-        // 1. 가장 많이 틀린 category의 ID와 상위 category 이름 조회
-        const mostIncorrectCategory = await ProblemModel.getMostIncorrectCategoryId(userId);
-        
-        if (!mostIncorrectCategory || mostIncorrectCategory.length === 0) {
-            return null;  // 해당 사용자가 틀린 문제가 없는 경우 처리
-        }
-
-        const { category_id, main_category, category } = mostIncorrectCategory[0];
-
-        // 2. 해당 category의 하위 카테고리 조회
-        const subCategories = await ProblemModel.getSubCategoriesByCategoryId(category_id, userId);
-
-        // 3. 최종 데이터 조합
-        return {
-            mainCategory: main_category,
-            category: category,
-            subCategories: subCategories.map(sub => ({
-                subCategory: sub.sub_category,
-                totalIncorrect: sub.total_incorrect
-            }))
-        };
-
-    } catch (error) {
-        throw new BaseError(status.INTERNAL_SERVER_ERROR, "틀린 문제 유형 통계 조회 실패");
-    }
-  },
-  
-
-  getStatisticIncorrectRatio: async (userId, categoryId) => {
-    try {
-        // 특정 category_id에 대한 하위 subcategory 비율 조회
-        const result = await ProblemModel.getIncorrectRatioByCategoryId(categoryId, userId);
-        return result;
-    } catch (error) {
-        throw new BaseError(status.INTERNAL_SERVER_ERROR, "틀린 문제 비율 통계 조회 실패");
-    }
-},
-
-
 };
